broke the screenGame HTML up into seperate components
